[Untested] Rewrite `insertItem` logic
That was surprisingly fun. The original code remains
as comments since my `:generateAssets` task is broken and triggers every time I call `:runClient` which is annoying as all hell.

This code is *untested* and could contain regressions. As far as I was aware all original mod logic remains intact but if you looked how this code was, it's very likely that you'd miss something. (a living example is `removeItem` btw, since the code looks exactly the same)
---
**This is a `roridev` and more importantly an `experiments` branch. For the love of god don't base your work in this branch. I will without further warning rebase and break history in ways that only the most patient will want to work with. Consider this a little peek into my inner ramblings and nothing more than that.**